+++ src/kits/interface/textview_support/StyleBuffer.h
@@ -56,7 +56,7 @@ class _BStyleRunDescBuffer_ : public _BTextViewSupportBuffer_<STEStyleRunDesc> {
 		void	InsertDesc(STEStyleRunDesc* inDesc, int32 index);
 		void	RemoveDescs(int32 index, int32 count = 1);
 
-		long	OffsetToRun(int32 offset) const;
+		int32	OffsetToRun(int32 offset) const;
 		void	BumpOffset(int32 delta, int32 index);
 
 		STEStyleRunDesc* operator[](int32 index) const;
